05.12.2021
-file 'tests' was replaced at 'fill_db'
-method 'create_tags' was replaced at 'update_tags'
-method 'create_todos' was update
-made new file 'constants'
