+++ todo/todo_app/constants.py
@@ -1,0 +1,12 @@
+all_tags = [
+    "ubuntu", "windows", "vmware", "vsphere", "redhat",
+    "debian", "warning", "bug", "degradation", "error",
+    "critical", "cloud", "cluster", "container", "jenkins",
+    "docker", "database", "python", "java", "go",
+    "webserver", "backend", "frontend",
+]
+# os = ["ubuntu", "windows", "vmware", "vsphere", "redhat", "debian"]
+# severity = ["warning", "bug", "degradation", "error", "critical"]
+# infrastructure = ["cloud", "cluster", "container"]
+# component = ["jenkins", "docker", "database", "python", "java", "go", "webserver", "backend", "frontend"]
+
